Prototype 6: Invulnerability, Knockdowns, Throw Animation
Added throws and the "knocked down" state, along with invulnerability to prevent "stun-locking".
Also added placeholder animations for throwing.

Originally, the changes were all minor in this patch, as larger changes were planned to be made when more art was produced.  But, this prototype is being uploaded as a backup of the original "engine", securing the original system in case the movement system rework is unsuccessful or becomes unnecessary.